chore: rename, add comment
